HandleFolderSelect component created
